refactor(context): move _run_git to github/_utils.py
